Try return - a new feature to work together with printing eval. If this feature is off then typing code works as it did before.
With this feature on:
If the code typed does parse with "return " concatenated at the
front then this version is used instead.
Else: the error of that first parse is ignored and the code
is parsed as is and if this second parse fails then error is
displayed.

Then, if either of the parses succeeded then the resulting
function is ran and error/result displayed.

This feature is ON by default.